put endpoint to a more reasonable place
